requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
numpy==1.21.3
opencv-python==4.5.4.58
psutil==5.8.0
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
simpleaudio==1.0.4
twilio==7.2.0
urllib3==1.26.7
